Find pg_dump if not in the provided path
Newer postgres packages put pg_dump in a different location and the
package naming scheme has changed slightly. So that we don't have to
change this with every version of postgres that comes out, we'll search
for pg_dump and add that to the path. Leaves the old path alone for
backwards compat although it is probably not needed.